pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hexlet-code"
version = "0.1.0"
description = "Second project - gendiff"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "WhiteA77", email = "second_project@easy.com"}]
dependencies = [
    "PyYAML>=6.0"
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.7.1",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
python_files = "test_*.py"

[tool.ruff]
line-length = 80
target-version = "py312"
select = ["E", "F"]
ignore = []
exclude = ["build", ".venv"]
